split distribution tests into 2 files
